dom/base/test/test_blobURL_expiring.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test for Blob URI expiration</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
  <script>

onmessage = function(e) {
  var blobURL = e.data;

  (new Promise(function(resolve, reject) {
    var xhr = new XMLHttpRequest();
    xhr.open("GET", blobURL);
    xhr.send();
    xhr.onload = function() {
      is(xhr.response, "123", "Response matches!");
      resolve();
    }
  })).then(function() {
    document.body.removeChild(iframe);
  }).then(function() {
    var xhr = new XMLHttpRequest();
    xhr.open("GET", blobURL);
    try {
      xhr.send();
      ok(false, "The URL should be done!");
    } catch(e) {
      ok(true, "The URL should be done!");
    }

    SimpleTest.finish();
  });
}

var iframe = document.createElement('iframe');
iframe.src = 'file_blobURL_expiring.html';
document.body.appendChild(iframe);

SimpleTest.waitForExplicitFinish();

  </script>
</body>
</html>
